requirements.txt
Django==5.1.3
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-rest-auth==4.0.1
django-allauth==0.50.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
isodate==0.7.2
lxml==5.3.0
Markdown==3.7
oauthlib==3.2.2
psycopg2-binary
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
python3-saml==1.16.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xmlsec==1.3.14
gunicorn
whitenoise